Corrected typo in author name
Dear

I am the corresponding (first) author of this paper and noticed that the surname of the last author, Luc De Raedt, contains a typo. Consequently, this publication has an incorrect citation entry and it is not linked as a publication of Luc De Raedt on Google Scholar and other platforms. 

Please remedy this typo such that this paper obtains a correct citation.

Kind regards,
Lennert De Smet